Updated - Moved guide and change log into rails folder
