+++ rails/client/app/bundles/kaiju/components/Guide/Guide.jsx
@@ -2,7 +2,7 @@ import React from 'react';
 import marked from 'marked';
 import classNames from 'classnames/bind';
 import 'github-markdown-css';
-import guide from '../../../../../../../docs/guide.md';
+import guide from '../../../../../../guide.md';
 import styles from './Guide.scss';
 
 const cx = classNames.bind(styles);
